Performance Optimization Test V1
- Reduce repeated model face-rule scans, scene metadata allocations and hit-group string copies.
- Skip unchanged sampler publication while preserving mipmap changes and resource retirement.
- Pack CPU entity output directly into owned staging buffers and add GPU conversion for eligible PBR geometry.
- Cache immutable metadata for published chunks while preserving per-frame transforms and motion history.
- Defer Vulkan raster-state commands outside active raster passes while preserving observable state and draw behavior.
- Share persistent model geometry and BLAS resources with bounded caches and submission-safe resource lifetimes.
- Correct ray-cone texture LOD for transformed instances across the affected path-tracing shaders.
- Add optional correlated CPU/GPU profiling, behavioral regression tests and bounded benchmark records.

+++ src/core/render/modules/world/ray_tracing/submodules/world_prepare.hpp
@@ -32,7 +32,7 @@ class WorldPrepare : public SharedObject<WorldPrepare> {
     void build();
 
   private:
-    using EntityRenderDataBatch = std::map<int, std::pair<std::shared_ptr<Entity>, VkTransformMatrixKHR>>;
+    using EntityRenderDataBatch = std::map<uint64_t, std::pair<std::shared_ptr<Entity>, VkTransformMatrixKHR>>;
     using ChunkTransformBatch =
         std::map<std::shared_ptr<Chunk1>, glm::dmat4, std::owner_less<std::shared_ptr<Chunk1>>>;
